fix t/leach.t for perls older than 5.14
The intent of the test is: look, I know that

   leach 2..8

isn't going to work, but that's okay because

   splice 2..8, 0, 0, 1

isn't going to work either. So in terms of "not working",
the best we can do is throw an error from 'leach LIST' that
looks a lot like the error from 'splice LIST'.

In order to figure out which error we need to expect from
'leach LIST', the test runs 'splice LIST'. On perl v5.14 and
beyond, this was trappable runtime error; on perl v5.12 and
lower, this is a compile time error.

Consequently, the testing strategy was completely broken for
perls older than 5.14.

One solution would have been to remove the test; but switching
from 'eval BLOCK' to 'eval STRING' is slightly better here,
since (a) it works, and (b) therefore we maintain test coverage.

Drive-by: version bump.