fix testcount on t/hash.t
fix warnings on tests that use t/lib/Test/Easy.pm as a
shim to provide backwards testing functionality: namely,
Test::More::subtest on perl 5.12.05 has a prototype ($&),
whereas earlier versions of T:M did not provide subtest
at all (hence bootstrapping ours in place did not trigger
prototype mismatch errors), and later versions of T:M:subtest
do not have a prototype.

This is another thing that was trivial with perlbrew, but
would be difficult without it.